fix menu base position, menu content refactore to const, add burger component
